Build lambda function separately
